+++ sdk/identity/identity/src/credentials/managedIdentityCredential/arcMsi.ts
@@ -21,7 +21,11 @@ const logger = credentialLogger(msiName);
 /**
  * Generates the options used on the request for an access token.
  */
-function prepareRequestOptions(scopes: string | string[]): PipelineRequestOptions {
+function prepareRequestOptions(
+  scopes: string | string[],
+  clientId?: string,
+  resourceId?: string
+): PipelineRequestOptions {
   const resource = mapScopesToResource(scopes);
   if (!resource) {
     throw new Error(`${msiName}: Multiple scopes are not supported.`);
@@ -31,12 +35,19 @@ function prepareRequestOptions(scopes: string | string[]): PipelineRequestOption
     "api-version": azureArcAPIVersion,
   };
 
-  const query = new URLSearchParams(queryParameters);
+  if (clientId) {
+    queryParameters.client_id = clientId;
+  }
+  if (resourceId) {
+    queryParameters.msi_res_id = resourceId;
+  }
 
   // This error should not bubble up, since we verify that this environment variable is defined in the isAvailable() method defined below.
   if (!process.env.IDENTITY_ENDPOINT) {
     throw new Error(`${msiName}: Missing environment variable: IDENTITY_ENDPOINT`);
   }
+
+  const query = new URLSearchParams(queryParameters);
 
   return createPipelineRequest({
     // Should be similar to: http://localhost:40342/metadata/identity/oauth2/token
@@ -96,13 +107,7 @@ async function filePathRequest(
  * Defines how to determine whether the Azure Arc MSI is available, and also how to retrieve a token from the Azure Arc MSI.
  */
 export const arcMsi: MSI = {
-  async isAvailable({ scopes, clientId }): Promise<boolean> {
-    if (clientId) {
-      logger.info(
-        `${msiName}: Unavailable. User assigned identity is not supported by the Azure Arc Managed Identity Endpoint. To authenticate with the system assigned identity, omit the client id when constructing the ManagedIdentityCredential, or if authenticating with the DefaultAzureCredential ensure the AZURE_CLIENT_ID environment variable is not set.`
-      );
-      return false;
-    }
+  async isAvailable({ scopes }): Promise<boolean> {
     const resource = mapScopesToResource(scopes);
     if (!resource) {
       logger.info(`${msiName}: Unavailable. Multiple scopes are not supported.`);
@@ -120,11 +125,16 @@ export const arcMsi: MSI = {
     configuration: MSIConfiguration,
     getTokenOptions: GetTokenOptions = {}
   ): Promise<AccessToken | null> {
-    const { identityClient, scopes, resourceId } = configuration;
+    const { identityClient, scopes, clientId, resourceId } = configuration;
 
+    if (clientId) {
+      logger.warning(
+        `${msiName}: user-assigned identities not supported. The argument clientId might be ignored by the service.`
+      );
+    }
     if (resourceId) {
       logger.warning(
-        `${msiName}: User defined managed Identity by resource Id is not supported by the Azure Arc Managed Identity Endpoint. Argument resourceId will be ignored.`
+        `${msiName}: user defined managed Identity by resource Id is not supported. Argument resourceId will be ignored.`
       );
     }
 
@@ -134,7 +144,7 @@ export const arcMsi: MSI = {
       disableJsonStringifyOnBody: true,
       deserializationMapper: undefined,
       abortSignal: getTokenOptions.abortSignal,
-      ...prepareRequestOptions(scopes),
+      ...prepareRequestOptions(scopes, clientId, resourceId),
       allowInsecureConnection: true,
     };